+++ src/test/java/com/project/socket/postrecruit/model/PostRecruitTest.java
@@ -1,17 +1,18 @@
-package com.project.socket.recruit;
+package com.project.socket.postrecruit.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.project.socket.postrecruit.model.PostRecruit;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 
 @DisplayNameGeneration(ReplaceUnderscores.class)
-class RecruitTest {
+class PostRecruitTest {
 
   @Test
-  void Recruit_엔티티_생성() {
-    Recruit recruit = Recruit.builder().id(1L).build();
-    assertThat(recruit).isNotNull();
+  void postRecruit_엔티티_생성() {
+    PostRecruit postRecruit = PostRecruit.builder().id(1L).build();
+    assertThat(postRecruit).isNotNull();
   }
 }
