+++ src/test/java/com/project/socket/skill/model/SkillTest.java
@@ -1,17 +1,18 @@
-package com.project.socket.recruitskill;
+package com.project.socket.skill.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.project.socket.skill.model.Skill;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 
 @DisplayNameGeneration(ReplaceUnderscores.class)
-class RecruitSkillTest {
+class SkillTest {
 
   @Test
-  void RecruitSkill_엔티티_생성() {
-    RecruitSkill recruitSkill = RecruitSkill.builder().id(1L).build();
-    assertThat(recruitSkill).isNotNull();
+  void Skill_엔티티_생성() {
+    Skill skill = Skill.builder().id(1L).build();
+    assertThat(skill).isNotNull();
   }
 }
